Read multiple ChannelVariables when they exist in AMI Event payload.
Channel variables specific to a channel can be conveyed in each AMI
event related to that channel. When this occurs, each variable is
referenced in a ChanVariable field. The value of a ChanVariable field
will always be of the form key=value, where key is the name of the
channel variable and value is its value.

Events that relate multiple channels will prefix these fields with an
event specific role specifier. For example, a DialBegin or a DialEnd
event will prefix the outbound channel's fields with Dest. So, the
Channel field is the DestChannel field; the Uniqueid field is the
DestUniqueid field, etc.

https://wiki.asterisk.org/wiki/display/AST/AMI+v2+Specification#AMIv2Specification-event_chanvariable
https://wiki.asterisk.org/wiki/display/AST/AMI+v2+Specification#AMIv2Specification-CommonFields